adds DataExporterStep class that extends the general ReleaseStep class
